Fix 64 to 32 bits casts warnings on msvc

+++ rzgamereconnect/src/AuthServer/AuthSession.cpp
@@ -113,7 +113,7 @@ void AuthSession::sendAccountList() {
 	synchronizedWithAuth = true;
 
 	TS_GA_ACCOUNT_LIST* accountListPacket;
-	const int maxCount = accountList.size() <= MAX_COUNT_PER_PACKET ? accountList.size() : MAX_COUNT_PER_PACKET;
+	const int maxCount = (int)(accountList.size() <= MAX_COUNT_PER_PACKET ? accountList.size() : MAX_COUNT_PER_PACKET);
 	accountListPacket = TS_MESSAGE_WNA::create<TS_GA_ACCOUNT_LIST, TS_GA_ACCOUNT_LIST::AccountInfo>(maxCount);
 
 	debug("Sending %d accounts\n", (int)accountList.size());
